+++ app/gui/src/project-view/bindings.ts
@@ -13,6 +13,7 @@ export const documentationEditorBindings = defineKeybinds('documentation-editor'
   toggle: ['Mod+D'],
   openLink: ['Mod+PointerMain'],
   paste: ['Mod+V'],
+  pasteRaw: ['Mod+Shift+V'],
 })
 
 export const interactionBindings = defineKeybinds('current-interaction', {