+++ CHANGELOG.md
@@ -29,6 +29,8 @@
 - ["Write" button in component menu allows to evaluate it separately from the
   rest of the workflow][11523].
 - [The documentation editor can now display tables][11564]
+- [The documentation editor supports the Markdown URL syntax, and uses it to
+  render pasted URLs as links][11597]
 - [Table Input Widget is now matched for Table.input method instead of
   Table.new. Values must be string literals, and their content is parsed to the
   suitable type][11612].
@@ -51,6 +53,7 @@
 [11547]: https://github.com/enso-org/enso/pull/11547
 [11523]: https://github.com/enso-org/enso/pull/11523
 [11564]: https://github.com/enso-org/enso/pull/11564
+[11597]: https://github.com/enso-org/enso/pull/11597
 [11612]: https://github.com/enso-org/enso/pull/11612
 
 #### Enso Standard Library